requirements-local.txt
pymol-open-source
imageio-ffmpeg
numpy